Made lottery draws accessible for each user separately (multiuser functionality)

+++ lottery/views.py
@@ -9,9 +9,6 @@ from models import Draw, User
 # CONFIG
 lottery_blueprint = Blueprint('lottery', __name__, template_folder='templates')
 
-user = User.query.first()
-draw_key = user.draw_key
-
 
 # VIEWS
 # View lottery page.
@@ -30,8 +27,7 @@ def add_draw():
     submitted_draw.strip()
 
     # Create a new draw with the form data.
-    new_draw = Draw(user_id=1, draw=submitted_draw, win=False, round=0, draw_key=draw_key)
-    # TODO: update user_id [user_id=1 is a placeholder]
+    new_draw = Draw(user_id=current_user.id, draw=submitted_draw, win=False, round=0, draw_key=current_user.draw_key)
 
     # Add the new draw to the database.
     db.session.add(new_draw)
@@ -47,7 +43,7 @@ def add_draw():
 @login_required
 def view_draws():
     # Get all draws that have not been played [played=0].
-    playable_draws = Draw.query.filter_by(played=False).all()  # TODO: filter playable draws for current user
+    playable_draws = Draw.query.filter_by(played=False, user_id=current_user.id).all()
 
     # If playable draws exist.
     if len(playable_draws) != 0:
@@ -60,7 +56,7 @@ def view_draws():
 
         # Decrypt each draw and add it to decrypted_draws array.
         for d in draw_copies:
-            d.view_draw(draw_key)
+            d.view_draw(current_user.draw_key)
             decrypted_draws.append(d)
 
         # Re-render lottery page with playable draws.
@@ -75,7 +71,7 @@ def view_draws():
 @login_required
 def check_draws():
     # Get played draws.
-    played_draws = Draw.query.filter_by(played=True).all()  # TODO: filter played draws for current user
+    played_draws = Draw.query.filter_by(played=True, user_id=current_user.id).all()
 
     # If played draws exist.
     if len(played_draws) != 0:
@@ -91,7 +87,7 @@ def check_draws():
 @lottery_blueprint.route('/play_again', methods=['POST'])
 @login_required
 def play_again():
-    delete_played = Draw.__table__.delete().where(Draw.played)  # TODO: delete played draws for current user only
+    delete_played = Draw.__table__.delete().where(Draw.played, user_id=current_user.id)
     db.session.execute(delete_played)
     db.session.commit()
